gateway/requirements_txt.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
requests==2.31.0
Werkzeug==3.0.1